Proposal to include arrays
Hi Luke,

This is my first pull request, I really hope I haven't messed anything up...

While browsing the UKB data showcase for other things, I got a better understanding of how variables are coded in relation to instances and arrays which led me to revisit the issue with the blood pressure variables we have discussed with Jie earlier. I discovered that each instance had two arrays and that these weren't included when updating the rap-variables.csv. It worked when I edited it manully, but I thought that something causes it to not update properly. And that led me to this script. I looked in the commit history and saw that arrays have been an issue before.

As I've commented in the file, the list of exceptions from removing arrays is longer than expected.

I guess for a start, maybe you can include the variables I have listed. But maybe all arrays should just be included? This of course may cause issues for other researchers using the ukbAid who have used variables where you have removed arrays if they download a new set of variables where arrays are included.

 What do you think?

Best, Niels